App(search): Search works as intended, for now
The search-container.vue component allows users to select
a species or water type to search by. Each input emits a selected
event. The parent component will respond with an http call to the
api.

The payload of the selected event are the parameter for the API call.
Currently, `s` which is for the search parameter. The search container
takes care of formatting for the API query string params.

The search container will render cards of the water bodies returned
by the API. The parent container `index.vue` has a computed prop
that contains the IDs for the fishable waters from the query.
This array of IDs is then used to filter the geojson geometry
of the fishable waters displayed by the `fishable-waters-map.vue`
component. Also a computed prop, so that it filters when the
search params change.

Using this method saves subsequent calls to the API, and database,
for filtering the geojson object. In other words, the geojson is
reactive to search input. Pretty cool!