App(feature): Add species pages/route to application
Each species in the database has it's own page with species details
a map of where to find the species, and a table of all the trophy
fish entries for that species.

These species pages represent one of the sections already within
the annual Trophy Fish Booklet.

Future work includes adding detailed descriptions of the species
and highlighting table rows based on trophy status.