Feat(geobuf): Fishable waters is requested as a geobuf
Huge performance gain on the request. Less data is transfered
and the timing is much faster. Decoding it in the App is fast.

Implementation of the geobuf route in the server needs work. For
simplicity, and in the interesting in prepping for the demo
tomorrow I'm going to wait to deploy as is if the build works.